+++ leads/data_persistence/analyzer/__init__.py
@@ -1,3 +1,3 @@
-from leads.data_persistence.analyzer.dp import *
+from leads.data_persistence.analyzer.dynamic import *
 from leads.data_persistence.analyzer.inference import *
-from leads.data_persistence.analyzer.processor import *
+from leads.data_persistence.analyzer.utils import *
