+++ leads/data_persistence/analyzer/static.py
@@ -1,12 +1,21 @@
+from importlib.util import find_spec as _find_spec
+
+if not _find_spec("matplotlib"):
+    raise ImportError("Please install `matplotlib` to run this module\n>>>pip install matplotlib")
+
 from datetime import datetime as _datetime
 from typing import Any as _Any, Callable as _Callable, Sequence as _Sequence
+from matplotlib.pyplot import figure as _figure, scatter as _scatter, show as _show, title as _title, \
+    colorbar as _colorbar, bar as _bar, xticks as _xticks, legend as _legend, xlabel as _xlabel, ylabel as _ylabel
 
 from leads.data import dlat2meters, dlon2meters, format_duration
+from leads.data_persistence.analyzer.utils import time_invalid, speed_invalid, mileage_invalid, latitude_invalid, \
+    longitude_invalid
 from leads.data_persistence.core import CSVDataset, DEFAULT_HEADER
 from .._computational import sqrt as _sqrt
 
 
-class Processor(object):
+class StaticProcessor(object):
     def __init__(self, dataset: CSVDataset) -> None:
         if DEFAULT_HEADER in dataset.read_header():
             raise KeyError("Your dataset must include the default header")
@@ -49,42 +58,6 @@ class Processor(object):
     def dataset(self) -> CSVDataset:
         return self._dataset
 
-    @staticmethod
-    def distance_between(lat_0: float, lon_0: float, lat: float, lon: float) -> float:
-        """
-        Calculate the distance between two locations on the Earth.
-        :param lat_0: the latitude of the first location
-        :param lon_0: the longitude of the first location
-        :param lat: the latitude of the second location
-        :param lon: the longitude of the second location
-        :return:
-        """
-        return _sqrt(dlon2meters(lon - lon_0, .5 * (lat_0 + lat)) ** 2 + dlat2meters(lat - lat_0) ** 2)
-
-    @staticmethod
-    def time_invalid(o: _Any) -> bool:
-        return not isinstance(o, int)
-
-    @staticmethod
-    def speed_invalid(o: _Any) -> bool:
-        return not isinstance(o, int | float) or o != o or o < 0
-
-    @staticmethod
-    def acceleration_invalid(o: _Any) -> bool:
-        return not isinstance(o, int | float) or o != o
-
-    @staticmethod
-    def mileage_invalid(o: _Any) -> bool:
-        return not isinstance(o, int | float) or o != o
-
-    @staticmethod
-    def latitude_invalid(o: _Any) -> bool:
-        return not isinstance(o, int | float) or o != o or not -90 < o < 90
-
-    @staticmethod
-    def longitude_invalid(o: _Any) -> bool:
-        return not isinstance(o, int | float) or o != o or not -180 < o < 180
-
     def bake(self) -> None:
         """
         Prepare the prerequisites for `process()`.
@@ -95,8 +68,8 @@ class Processor(object):
             t = int(row["t"])
             speed = row["speed"]
             mileage = row["mileage"]
-            if Processor.time_invalid(t) or Processor.speed_invalid(
-                    speed) or Processor.mileage_invalid(mileage):
+            if time_invalid(t) or speed_invalid(
+                    speed) or mileage_invalid(mileage):
                 self._invalid_rows.append(i)
                 return
             if self._start_time is None:
@@ -111,7 +84,7 @@ class Processor(object):
             self._end_mileage = mileage
             lat = row["latitude"]
             lon = row["longitude"]
-            if not row["gps_valid"] or Processor.latitude_invalid(lat) or Processor.longitude_invalid(lon):
+            if not row["gps_valid"] or latitude_invalid(lat) or longitude_invalid(lon):
                 self._gps_invalid_rows.append(i)
             else:
                 if self._min_lat is None or lat < self._min_lat:
@@ -146,7 +119,7 @@ class Processor(object):
         return (
             f"Baked {self._valid_rows_count} / {self._read_rows_count} ROWS",
             f"Baking Rate: {100 * self._valid_rows_count / self._read_rows_count:.2f}%",
-            f"Skipped Rows: {Processor._hide_others(self._invalid_rows, 5)}",
+            f"Skipped Rows: {StaticProcessor._hide_others(self._invalid_rows, 5)}",
             f"Start Time: {_datetime.fromtimestamp(self._start_time * .001).strftime("%Y-%m-%d %H:%M:%S")}",
             f"End Time: {_datetime.fromtimestamp(self._end_time * .001).strftime("%Y-%m-%d %H:%M:%S")}",
             f"Duration: {format_duration(self._duration * .001)}",
@@ -155,7 +128,7 @@ class Processor(object):
             f"v\u2098\u2090\u2093: {self._max_speed:.2f} KM / H",
             f"v\u2090\u1D65\u1D4D: {self._avg_speed:.2f} KM / H",
             f"GPS Hit Rate: {100 * self._gps_valid_count / self._valid_rows_count:.2f}%",
-            f"GPS Skipped Rows: {Processor._hide_others(self._gps_invalid_rows, 5)}"
+            f"GPS Skipped Rows: {StaticProcessor._hide_others(self._gps_invalid_rows, 5)}"
         )
 
     def erase_unit_cache(self) -> None:
@@ -261,3 +234,71 @@ class Processor(object):
 
     def close(self) -> None:
         self._dataset.close()
+
+    def draw_lap(self, lap_index: int = -1) -> None:
+        if lap_index < 0:
+            for i in range(len(self._laps)):
+                self.draw_lap(i)
+        if lap_index >= len(self._laps):
+            raise IndexError("Lap index out of range")
+
+        def unit(row: dict[str, _Any], index: int) -> None:
+            if index < (lap := self._laps[lap_index])[0] or index > lap[1]:
+                return
+            t = int(row["t"])
+            if self._lap_start_time is None:
+                self._lap_start_time = t
+            self._lap_end_time = t
+            lat = row["latitude"]
+            lon = row["longitude"]
+            self._lap_d.append(row["speed"])
+            self._lap_x.append(x := dlon2meters(lon - self._min_lon, lat))
+            self._lap_y.append(y := dlat2meters(lat - self._min_lat))
+            if self._max_lap_x is None or x > self._max_lap_x:
+                self._max_lap_x = x
+            if self._max_lap_y is None or y > self._max_lap_y:
+                self._max_lap_y = y
+
+        self.foreach(unit, True, True)
+        far = max(self._max_lap_x, self._max_lap_y)
+        self._lap_x.append(far)
+        self._lap_y.append(far)
+        self._lap_d.append(self._max_speed)
+        _figure(figsize=(6, 5))
+        _title(f"Lap {lap_index + 1} ({self._laps[lap_index][3]:.2f} KM @ {format_duration(
+            self._laps[lap_index][2] * .001)})")
+        _scatter(self._lap_x, self._lap_y, c=self._lap_d, cmap="hot_r")
+        _xlabel("X (M)")
+        _ylabel("Y (M)")
+        cb = _colorbar()
+        cb.set_label("Speed (KM / H)")
+        cb.ax.hlines(self._laps[lap_index][4], 0, 1)
+        _show()
+
+    def draw_comparison_of_laps(self, width: float = .3) -> None:
+        durations = []
+        x0 = []
+        distances = []
+        x1 = []
+        avg_speeds = []
+        x2 = []
+        x_ticks = []
+        i = 1
+        for lap in self._laps:
+            durations.append(lap[2] / self._max_lap_duration)
+            x0.append(i)
+            distances.append(lap[3] / self._max_lap_distance)
+            x1.append(i + width)
+            avg_speeds.append(lap[4] / self._max_lap_avg_speed)
+            x2.append(i + 2 * width)
+            x_ticks.append(f"L{i}")
+            i += 1
+        _figure(figsize=(5 * _sqrt(len(self._laps)), 5))
+        _bar(x0, durations, width, label="Duration")
+        _bar(x1, distances, width, label="Distance")
+        _bar(x2, avg_speeds, width, label="Average Speed")
+        _xticks(x1, x_ticks)
+        _legend()
+        _xlabel("Lap")
+        _ylabel("Proportion (% / max)")
+        _show()
